+++ client/src/Pages/Disaster.js
@@ -1,21 +1,34 @@
 import React, { useState, useEffect } from "react";
+import Table from "@mui/material/Table";
+import TableBody from "@mui/material/TableBody";
+import TableCell from "@mui/material/TableCell";
+import TableContainer from "@mui/material/TableContainer";
+import TableHead from "@mui/material/TableHead";
+import TableRow from "@mui/material/TableRow";
+import Paper from "@mui/material/Paper";
 import axios from "axios";
 import { useNavigate } from "react-router-dom";
 import Button from "@mui/material/Button";
 
-function Disaster() {
-  const navigate = useNavigate();
-  const [disasters, setDisasters] = useState([]);
+
+function CreatePost() {
+    const navigate = useNavigate();
+    const [disasters, setDisasters] = useState([]);
 
   useEffect(() => {
     loadDisasters();
-  }, []);
+  },[]);
 
   async function loadDisasters() {
-    const result = await axios.get(
-      'http://localhost:3001/disasters/'
-    );
-    setDisasters(result.data);
+    const result = await axios
+    .get('http://localhost:3001/disasters/')
+    .then((result) => {
+      setDisasters(result.data);
+      console.log("Result", disasters);
+    })
+    .catch((err) => {
+      console.log(err);
+    });
   }
 
   // Function to delete a disaster
@@ -25,47 +38,33 @@ function Disaster() {
     );
     loadDisasters();
   };
-
-  return (
-    <div className="home-page">
-      <table class="table">
-        <thead class="thead-dark">
-          <tr>
-            <th scope="col">#</th>
-            <th scope="col">Disaster</th>
-            <th scope="col">Date of Occurrence</th>
-            <th scope="col">Location</th>
-            <th>Action</th>
-          </tr>
-        </thead>
-        <tbody>
+  
+return (
+  
+  <TableContainer component={Paper}>
+    <Button variant="contained" onClick={() => navigate("/create")}>Add Event</Button>
+    <Table sx={{ minWidth: 150 }} aria-label="simple table" stickyHeader>
+      <TableHead>
+        <TableRow>
+          <TableCell>ID</TableCell>
+          <TableCell align="right">Event</TableCell>
+          <TableCell align="right">Date of Occurrence</TableCell>
+          <TableCell align="right">Location</TableCell>
+        </TableRow>
+      </TableHead>
+      <TableBody>
           {disasters.map((disaster, index) => (
-            <tr>
-              <th scope="row">{index + 1}</th>
-              <td>{disaster.disaster_name}</td>
-              <td>{disaster.disaster_date}</td>
-              <td>{disaster.disaster_location}</td>
-              <Button
-                variant="contained"
-              >
-                View
-              </Button>
-              
-              <Button
-                variant="contained"
-                color="primary"
-              >
-                Edit
-              </Button>
-              <Button variant="contained" color="secondary" onClick={() => deleteDisaster(disaster.id)}>
-                Delete
-              </Button>
-            </tr>
+            <TableRow key={disaster.id}>
+              <TableCell component="th" scope="row">{disaster.disaster_id}</TableCell>
+              <TableCell align="right">{disaster.disaster_name}</TableCell>
+              <TableCell align="right">{disaster.disaster_date}</TableCell>
+              <TableCell align="right">{disaster.disaster_location}</TableCell>
+            </TableRow>
           ))}
-        </tbody>
-      </table>
-    </div>
-  );
+          </TableBody>
+          </Table>
+          </TableContainer>
+          );
 }
 
-export default Disaster;
+export default CreatePost;
