+++ client/src/Permissions/Permissions.js
@@ -1,0 +1,5 @@
+const PERMISSIONS = {
+    CAN_VIEW_DISASTER: "view_disaster",
+};
+
+export default PERMISSIONS;
